refactor(ui): modularize jobs page sections, add shared content modules, jobs hooks, and supporting auth/social client structure
