+++ frontend/src/components/content/Navbar.content.mdx
@@ -1,0 +1,11 @@
+export const NAV_PRIMARY_ITEMS = [
+  { to: "/", label: "Home", icon: "home" },
+  { to: "/jobs", label: "Jobs", icon: "jobs" },
+  { to: "/tracker", label: "Tracker", icon: "tracker" },
+  { to: "/learn-more", label: "Learn More", icon: "learn" },
+];
+
+export const NAV_AUTH_LINKS = [
+  { to: "/sign-up", label: "Join" },
+  { to: "/sign-in", label: "Sign in" },
+];
